requirements.txt
openai==1.97.1
pydantic==2.11.7
python-dotenv==1.1.1
pypdf==5.9.0
polars==1.31.0
PyMuPDF==1.26.3
